+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Chessboard.js für Contao 4
 
+## Version 0.1.0 (2016-06-22)
+
+* Bei Folgezug, wo Ausgangs- und Zielfeld gleich sind, wurde versehentlich das Ausgangsfeld gelöscht
+* Nächster-Zug-Button: "Start" statt "Grundstellung"
+* Bei Folgezügen können zwei zusätzliche Felderfarben je Halbzug angegeben werden
+* Grundstellung eingebaut als Startstellung
+* Spielmodus-Auswahl über Radio-Buttons
+* Checkbox für Folgezüge mit Halbzug-Eingabe
+* Integration FEN-Folgezüge mit Markierungen
+
 ## Version 0.0.3 (2016-06-18)
 
 * Abhängigkeit menatwork/contao-multicolumnwizard-bundle hinzugefügt